requirements.txt
akshare>=1.12.0
pandas>=2.0.0
jinja2>=3.1.0
pyyaml>=6.0
requests>=2.31.0
yfinance>=0.2.0
